+++ xcube/cli/gen.py
@@ -67,11 +67,11 @@ resampling_algs = NAME_TO_GDAL_RESAMPLE_ALG.keys()
                    f'The choices for the resampling algorithm are: {resampling_algs}')
 @click.option('--append', '-a', is_flag=True,
               help='Append successive outputs.')
-@click.option('--sort', is_flag=True,
-              help='This parameter is only needed for creating netCDF cubes. '
-                   'The input file list will be sorted before creating the data cube. '
-                   'If --sort parameter is not passed, order of input list will be kept.'
-                   'When creating zarr data cubes, chronology is forced by "xcube gen". ')
+@click.option('--nosort', is_flag=True,
+              help='This parameter is needed if the order of the input files is already sorted by the user and the '
+                   'chronology is ensured. By providing a sorted list of inputs the performance of xcube gen can be '
+                   'improved. When the parameter is not given, the default behaviour of xcube gen sorts the input files'
+                   ' based on the time stamps of the input files.')
 @click.option('--info', '-i', is_flag=True,
               help='Displays additional information about format options or about input processors.')
 @click.option('--dry_run', is_flag=True,
@@ -88,7 +88,7 @@ def gen(inputs: str,
         append: bool,
         dry_run: bool,
         info: bool,
-        sort: bool):
+        nosort: bool):
     """
     Generate data cube.
     Data cubes may be created in one go or successively in append mode, input by input.
@@ -108,7 +108,7 @@ def gen(inputs: str,
     append_mode = append
     dry_run = dry_run
     info_mode = info
-    sort_mode = sort
+    no_sort = nosort
 
     # Force loading of plugins
     __import__('xcube.util.plugin')